Port application to react.js.
The component philosophy makes a lot of sense to me since my current hiccups in writing this app are dealing with the constantly changing UI and consequently having to constantly pay attention to the coupling between DOM, event bindings, and the app. So far the components encapsulate related logic better.

My hope is to improve iteration speed.